+++ src/file-io/fits-loading.jl
@@ -1,22 +1,8 @@
 export load_spectral_dataset
 
-function load_spectral_dataset(
-    grp_path,
-    rm_path;
-    mission::AbstractMissionTrait = NoAssociatedMission(),
-    T::Type = Float64,
-)
-    f_grp = FITS(grp_path)
-    f_rm = FITS(rm_path)
-
-    (counts, countserror, channels) = parse_grp_fits_file(mission, f_grp)
+function load_response_matrix(fits, mission, T)
     (matrix, rm_channels, rm_low_energy_bins, rm_high_energy_bins) =
-        parse_rm_fits_file(mission, f_rm)
-
-    meta = meta_from_fits(mission, f_grp, f_rm, grp_path, rm_path)
-
-    close(f_grp)
-    close(f_rm)
+        parse_rm_fits_file(mission, fits, T)
 
     rm::ResponseMatrix{T} = ResponseMatrix(
         matrix,
@@ -25,7 +11,12 @@ function load_spectral_dataset(
         T.(rm_high_energy_bins),
     )
 
+    return rm
+end
+
+function make_spectral_dataset(meta::AbstractSpectralDatasetMeta, rm::ResponseMatrix, counts, countserror, channels, T)
     low_energy_bins, high_energy_bins = augmented_energy_channels(channels, rm)
+
     counts_vec::Vector{T} = T.(counts)
     countserr_vec::Vector{T} = T.(countserror)
     channels_vec::Vector{Int} = Int.(channels)
@@ -41,3 +32,5 @@ function load_spectral_dataset(
         false,
     )
 end
+
+load_spectral_dataset(::AbstractMissionTrait, args...; kwargs...) = error("First argument must be a mission trait.")
