+++ src/file-io/missions/abstract-mission.jl
@@ -1,0 +1,57 @@
+export AbstractMissionTrait,
+    NoAssociatedMission,
+    missiontrait,
+    drop_bad_channels!
+
+# for future use: mission specific parsing
+abstract type AbstractMissionTrait end
+struct NoAssociatedMission <: AbstractMissionTrait end
+
+missiontrait(T::Type{<:AbstractSpectralDatasetMeta}) = error("No trait set for $(T)")
+missiontrait(::M) where {M<:AbstractSpectralDatasetMeta} = missiontrait(M)
+missiontrait(::SpectralDataset{M}) where {M} = missiontrait(M)
+
+# interface depending on the meta
+drop_bad_channels!(::SpectralDataset, M::AbstractMissionTrait; kwargs...) =
+    nothing
+drop_bad_channels!(sd::D; kwargs...) where {D<:SpectralDataset} = drop_bad_channels!(sd, missiontrait(D); kwargs...)
+drop_bad_channels!(::SpectralDataset{<:NoAssociatedMission}) = nothing
+
+# must return these four things
+function parse_rm_fits_file(mission::AbstractMissionTrait, fits, T)
+    rm_low_energy = read(fits[2], "ENERG_LO")
+    rm_high_energy = read(fits[2], "ENERG_HI")
+    F_chan = eachcol(read(fits[2], "F_CHAN"))
+    N_chan = eachcol(read(fits[2], "N_CHAN"))
+    matrix_lookup = read(fits[2], "MATRIX")
+
+    low_energy_bins = read(fits[3], "E_MIN")
+    high_energy_bins = read(fits[3], "E_MAX")
+    channels = read(fits[3], "CHANNEL")
+
+    n = length(channels)
+
+    offset = mission == NoAssociatedMission() ? 1 : 0
+    matrix = spzeros(T, n, n)
+
+    build_matrix_response!(
+        matrix,
+        # from rm table
+        rm_low_energy,
+        rm_high_energy,
+        F_chan,
+        N_chan,
+        matrix_lookup,
+        # from energy table
+        low_energy_bins,
+        high_energy_bins,
+        channels,
+        offset
+    )
+
+    (matrix, channels, low_energy_bins, high_energy_bins)
+end
+
+# include missions
+include("no-associated-mission.jl")
+include("xmm-newton-mission.jl")
